// src/interfaces/IClientService.ts
export interface IClientService<T> {
  collection: () => Promise<T[] | void>;
  model: (id: string) => Promise<T | void>;
  create: (data: object) => Promise<T | void>;
  update: (data: object) => Promise<T | void>;
  remove: (id: string) => Promise<T | void>;
  onChannel: () => Promise<void>;
  offChannel: () => Promise<void>;
}